# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ventus In-App Feedback API
  version: 0.1.0
  description: Versioned contract for browser ingestion, human workflows, and agent workflows.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /feedback:
    post:
      operationId: createFeedback
      security:
        - projectKey: []
        - bearerAuth: [feedback:submit]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFeedbackRequest" }
      responses:
        "201":
          description: Feedback accepted
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateFeedbackResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
    get:
      operationId: listFeedback
      security: [{ bearerAuth: [feedback:read] }]
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: projectId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: array, items: { $ref: "#/components/schemas/FeedbackStatus" } } }
        - { name: category, in: query, schema: { type: array, items: { $ref: "#/components/schemas/FeedbackCategory" } } }
        - { name: priority, in: query, schema: { type: array, items: { $ref: "#/components/schemas/FeedbackPriority" } } }
        - { name: label, in: query, schema: { type: array, items: { type: string } } }
        - { name: claimedBy, in: query, schema: { type: string } }
        - { name: release, in: query, schema: { type: string } }
        - { name: environment, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string, maxLength: 256 } }
        - { name: sort, in: query, schema: { enum: [created_at, updated_at, priority], default: created_at } }
        - { name: order, in: query, schema: { enum: [asc, desc], default: desc } }
      responses:
        "200":
          description: Cursor page
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Feedback" } }
                  nextCursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Error" }
  /feedback/{feedbackId}:
    parameters:
      - $ref: "#/components/parameters/FeedbackId"
    get:
      operationId: getFeedback
      security:
        - bearerAuth: [feedback:read]
        - reporterToken: []
      responses:
        "200":
          description: Feedback record
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateFeedback
      security: [{ bearerAuth: [feedback:triage] }]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateFeedbackRequest" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
        "409": { $ref: "#/components/responses/Error" }
  /feedback/{feedbackId}/comments:
    post:
      operationId: addFeedbackComment
      security: [{ bearerAuth: [feedback:comment] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, body]
              properties:
                version: { type: integer, minimum: 1 }
                body: { type: string, minLength: 1, maxLength: 10000 }
      responses:
        "201": { description: Comment added }
  /feedback/{feedbackId}/claim:
    post:
      operationId: claimFeedback
      security: [{ bearerAuth: [feedback:triage] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/LeaseMutation" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      operationId: releaseFeedbackClaim
      security: [{ bearerAuth: [feedback:triage] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/VersionMutation" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
  /feedback/{feedbackId}/claim/renew:
    post:
      operationId: renewFeedbackClaim
      security: [{ bearerAuth: [feedback:triage] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/LeaseMutation" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
        "409": { $ref: "#/components/responses/Error" }
  /feedback/{feedbackId}/resolve:
    post:
      operationId: resolveFeedback
      security: [{ bearerAuth: [feedback:resolve] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveFeedbackRequest" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
        "409": { $ref: "#/components/responses/Error" }
  /feedback/{feedbackId}/close:
    post:
      operationId: closeFeedback
      security: [{ bearerAuth: [feedback:close] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/NotedMutation" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
  /feedback/{feedbackId}/reopen:
    post:
      operationId: reopenFeedback
      security:
        - bearerAuth: [feedback:triage]
        - reporterToken: []
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/NotedMutation" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
  /feedback/{feedbackId}/reject:
    post:
      operationId: rejectFeedback
      security: [{ bearerAuth: [feedback:triage] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RejectFeedbackRequest" }
      responses:
        "200": { $ref: "#/components/responses/Feedback" }
  /feedback/{feedbackId}/evidence:
    post:
      operationId: addFeedbackEvidence
      security: [{ bearerAuth: [feedback:comment] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody: { $ref: "#/components/requestBodies/NotedMutation" }
      responses:
        "201": { description: Evidence appended }
  /feedback/{feedbackId}/events:
    get:
      operationId: listFeedbackEvents
      security: [{ bearerAuth: [feedback:read] }]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200": { description: Append-only event page }
  /feedback/{feedbackId}/attachments:
    post:
      operationId: uploadFeedbackAttachment
      security:
        - projectKey: []
        - bearerAuth: [feedback:comment]
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                kind: { enum: [screenshot, image, text, video, other] }
      responses:
        "201": { description: Attachment accepted }
        "413": { $ref: "#/components/responses/Error" }
  /feedback/{feedbackId}/attachments/{attachmentId}:
    get:
      operationId: downloadFeedbackAttachment
      security:
        - bearerAuth: [feedback:read]
        - reporterToken: []
      parameters:
        - $ref: "#/components/parameters/FeedbackId"
        - { name: attachmentId, in: path, required: true, schema: { type: string } }
      responses:
        "302": { description: Short-lived authorized object-storage URL }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    projectKey: { type: apiKey, in: header, name: X-Feedback-Project-Key }
    reporterToken: { type: apiKey, in: header, name: X-Feedback-Reporter-Token }
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT-or-service-token }
  parameters:
    FeedbackId: { name: feedbackId, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 200 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string }, description: Current feedback version as an ETag }
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
...
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Feedback:
      description: Updated feedback
      headers:
        ETag: { schema: { type: string } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FeedbackMutationResponse" }
  schemas:
    FeedbackCategory: { type: string, enum: [bug, feedback, idea] }
    FeedbackStatus: { type: string, enum: [new, triaged, in_progress, resolved, closed, rejected, reopened] }
    FeedbackPriority: { type: string, enum: [unset, low, medium, high, urgent] }
...
    CreateFeedbackRequest:
      type: object
      additionalProperties: false
      required: [schemaVersion, projectId, category, title, description]
      properties:
        schemaVersion: { const: "1.0" }
        projectId: { type: string, maxLength: 128 }
        category: { $ref: "#/components/schemas/FeedbackCategory" }
        title: { type: string, minLength: 1, maxLength: 140 }
        description: { type: string, minLength: 1, maxLength: 5000 }
        sourceApp: { type: string, maxLength: 128 }
        release: { type: string, maxLength: 128 }
        environment: { type: string, maxLength: 128 }
        url: { type: string, maxLength: 2048 }
        context: {}
        diagnostics: {}
        reporterTokenRequested: { type: boolean, default: false }
    CreateFeedbackResponse:
      type: object
      required: [feedback]
      properties:
        feedback: { $ref: "#/components/schemas/Feedback" }
        reporterToken: { type: string, description: Returned only once and stored hashed }
...